Fixed a bug related to the configuration file reading.
This commit fixes a bug that appeared when including a project that itself depends on Blammo. The blammo.xml file was always looked up in './blammo.xml', regardless of the real directory tree and where Blammo was used. Using the  __directory variable fixes this behavior.

Example:
my_project
    └── some_project_that_uses_blammo
        ├── lib
        ├── node_modules
        │   ├── blammo
        ├── blammo.xml

In this example, blammo.xml wasn't found (as the code referred to './', which was 'my_project/').